finsight/backend/requirements.txt
fastapi
uvicorn
langchain-groq
langchain-chroma
langchain-community
sentence-transformers
chromadb
mlflow
yfinance
beautifulsoup4
lxml
httpx
python-dotenv
pydantic
sse-starlette
langgraph
